extract global admin user model to it's own file
